% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_complement.R
\name{dna_complement}
\alias{dna_complement}
\title{Creates complement of DNA sequence (as vector)}
\usage{
dna_complement(chrstring)
}
\arguments{
\item{chrstring}{character vector describing DNA sequence}
}
\value{
Returns character vector of DNA sequence complement
}
\description{
Creates complement of DNA sequence (as vector)
}
\examples{
dna_complement(c("g", "a", "t", "t", "a", "c", "a"))
}
