% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_frame.R
\name{dna_frame}
\alias{dna_frame}
\title{Create DNA frame list object}
\usage{
dna_frame(charstring)
}
\arguments{
\item{charstring}{concatenated string of DNA letters}
}
\value{
DNA frame list object (not currently defined)
}
\description{
Create DNA frame list object
}
\examples{
dna_frame("gattaca")
}
