% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnagraph_diagnostic.R
\name{dnagraph_diagnostic}
\alias{dnagraph_diagnostic}
\title{Diagnose standard size requirements for A5 page from sequence}
\usage{
dnagraph_diagnostic(
  dnaobj,
  rowcount = max_rowlength(nrow(dnaobj$seqdata)),
  size = 5 * 50/rowcount,
  sepr = 0.007 * size,
  family = "DNAMosaic"
)
}
\arguments{
\item{dnaobj}{DNA or LAMP frame list object}

\item{rowcount}{Number of bases per row}

\item{size}{Font size to print bases}

\item{sepr}{Vertical separation between bases}

\item{family}{Font family; default is the custom 'DNAMosaic'}
}
\value{
returns list of key graph parameter settings under 'default' values
}
\description{
Diagnose standard size requirements for A5 page from sequence
}
\examples{
lframe <- dna_frame(random_sequence_generator(500))
dnagraph_diagnostic(lframe)
}
