% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnagraph_setup.R
\name{dnagraph_setup}
\alias{dnagraph_setup}
\title{Setup system for custom DNA graphing}
\usage{
dnagraph_setup(addfont = TRUE, showtext = TRUE)
}
\arguments{
\item{addfont}{T/F. Choose to add DNAMosaic custom font to system}

\item{showtext}{F/F. Choose to deploy showtext}
}
\value{
invisible
}
\description{
Setup system for custom DNA graphing
}
