% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp_dnagraph_single.R
\name{lamp_dnagraph_single}
\alias{lamp_dnagraph_single}
\title{Draw ssDNA of LAMP sequence and primers to fill A5 scale from LAMP frame object}
\usage{
lamp_dnagraph_single(
  dnaobj,
  rowcount = max_rowlength(nrow(dnaobj$seqdata)),
  size = 5 * 50/rowcount,
  sepr = 0.007 * size,
  family = "DNAMosaic"
)
}
\arguments{
\item{dnaobj}{LAMP frame list object}

\item{rowcount}{Number of bases per row}

\item{size}{Font size to print bases}

\item{sepr}{Vertical separation between bases}

\item{family}{Font family; default is the custom 'DNAMosaic'}
}
\value{
returns ggplot object scaled to A5 dimensions with DNA sequence and LAMP primers
}
\description{
Draw ssDNA of LAMP sequence and primers to fill A5 scale from LAMP frame object
}
\examples{
lframe <- lamp_frame("gattaca", f3 = "gatt", b3 = "tgta")
lamp_dnagraph_single(lframe, family = "serif")
}
