% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp_frame.R
\name{lamp_frame}
\alias{lamp_frame}
\title{Create LAMP Frame DNA object list}
\usage{
lamp_frame(
  dnastring,
  f3 = NULL,
  f2 = NULL,
  f1c = NULL,
  b1c = NULL,
  b2 = NULL,
  b3 = NULL,
  lf = NULL,
  lb = NULL
)
}
\arguments{
\item{dnastring}{concatenated string of DNA letters}

\item{f3}{charstring - F3 primer}

\item{f2}{charstring - F2 primer}

\item{f1c}{charstring - F1c primer}

\item{b1c}{charstring - B1c primer}

\item{b2}{charstring - B2 primer}

\item{b3}{charstring - B3 primer}

\item{lf}{charstring - LF primer}

\item{lb}{charstring - LB primer}
}
\value{
LAMP Frame list object
}
\description{
Create LAMP Frame DNA object list
}
\examples{
lamp_frame("gattaca", f3 = "gatt", b3 = "tgta")
}
