% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_rowlength.R
\name{max_rowlength}
\alias{max_rowlength}
\title{Max rowlength}
\usage{
max_rowlength(n)
}
\arguments{
\item{n}{Number of bases in DNA sequence}
}
\value{
Returns integer number of bases in a DNA row (currently under A5 page assumptions)
}
\description{
Max rowlength
}
\examples{
max_rowlength(100)
}
