% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_load_primerinfo.R
\name{pe_load_primerinfo}
\alias{pe_load_primerinfo}
\title{Load PrimerInfo file from Primer Explorer v5}
\usage{
pe_load_primerinfo(filename)
}
\arguments{
\item{filename}{file location of  PrimerInfo file}
}
\value{
returns LAMP frame object with 'core' primers attached
}
\description{
Load PrimerInfo file from Primer Explorer v5
}
