% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_add.R
\name{primer_add}
\alias{primer_add}
\title{Add primer to DNA frame object}
\usage{
primer_add(seqframe, primr, matchtype = "f")
}
\arguments{
\item{seqframe}{DNA frame list object}

\item{primr}{String sequence of Primer}

\item{matchtype}{Matchtype 'f', 'r', 'c', or 'rc'. Must be exact.}
}
\value{
Returns seqframe list with added primers.
}
\description{
Add primer to DNA frame object
}
\examples{
dframe <- dna_frame("gattaca")
primer_add(dframe, "atta", "f")
}
