% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_match.R
\name{primer_match}
\alias{primer_match}
\title{Match primer with DNA frame object}
\usage{
primer_match(seqframe, primr, matchtype = "f")
}
\arguments{
\item{seqframe}{DNA frame object}

\item{primr}{sequence to match char}

\item{matchtype}{"f", "r", "rc" or "c"}
}
\value{
Returns integer vector of sequence match locations
}
\description{
Match primer with DNA frame object
}
\examples{
dframe <- dna_frame("gattaca")
primer_match(dframe, "atta", "f")
}
