% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sequence_generator.R
\name{random_sequence_generator}
\alias{random_sequence_generator}
\title{Generate random DNA sequence of length n}
\usage{
random_sequence_generator(n)
}
\arguments{
\item{n}{Length of sequence to generate (integer)}
}
\value{
Charstring of length n with random DNA letters (including nonstandard)
}
\description{
Generate random DNA sequence of length n
}
\examples{
random_sequence_generator(10)
}
