% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dnagraph_double.R
\name{simple_dnagraph_double}
\alias{simple_dnagraph_double}
\title{Draw dsDNA to fill A5 scale from DNA frame object}
\usage{
simple_dnagraph_double(
  dnaobj,
  rowcount = max_rowlength(nrow(dnaobj$seqdata)),
  size = 5 * 50/rowcount,
  sepr = 0.035 * 50/rowcount,
  family = "DNAMosaic",
  upper = FALSE
)
}
\arguments{
\item{dnaobj}{DNA frame list object}

\item{rowcount}{Number of bases per row}

\item{size}{Font size to print bases}

\item{sepr}{Vertical separation between bases}

\item{family}{Font family; default is the custom 'DNAMosaic'}

\item{upper}{Whether to print DNA in UPPERCASE (default FALSE)}
}
\value{
returns ggplot object scaled to A5 dimensions with DNA sequence drawn.
}
\description{
Draw dsDNA to fill A5 scale from DNA frame object
}
\examples{
dframe <- dna_frame("gattaca")
simple_dnagraph_double(dframe, family = "serif")
}
