% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dnagraph_single.R
\name{simple_dnagraph_single}
\alias{simple_dnagraph_single}
\title{Draw ssDNA to fill A5 scale from DNA frame object}
\usage{
simple_dnagraph_single(
  dnaobj,
  rowcount = max_rowlength(nrow(dnaobj$seqdata)),
  size = 5 * 50/rowcount,
  sepr = 0.035 * 50/rowcount,
  family = "serif"
)
}
\arguments{
\item{dnaobj}{DNA frame list object}

\item{rowcount}{Number of bases per row}

\item{size}{Font size to print bases}

\item{sepr}{Vertical separation between bases}

\item{family}{Font family; default is the custom 'DNAMosaic'}
}
\value{
returns ggplot object scaled to A5 dimensions with DNA sequence drawn.
}
\description{
Draw ssDNA to fill A5 scale from DNA frame object
}
\examples{
dframe <- dna_frame("gattaca")
simple_dnagraph_single(dframe, family = "serif")
}
